+++ packages/zowe-explorer/__tests__/__unit__/dataset/dsNodeActions.unit.test.ts
@@ -16,6 +16,7 @@ import { ZoweDatasetNode } from "../../../src/dataset/ZoweDatasetNode";
 import * as brtimperative from "@zowe/imperative";
 import * as zowe from "@zowe/cli";
 import * as dsNodeActions from "../../../src/dataset/actions";
+import * as refreshActions from "../../../src/shared/refresh";
 import { Profiles } from "../../../src/Profiles";
 import { FAVORITE_CONTEXT, DS_SESSION_CONTEXT, FAV_SUFFIX } from "../../../src/globals";
 
@@ -219,8 +220,8 @@ describe("dsNodeActions", () => {
                     };
                 }),
             });
-            const spy = jest.spyOn(dsNodeActions, "refreshAll");
-            dsNodeActions.refreshAll(testDSTree);
+            const spy = jest.spyOn(refreshActions, "refreshAll");
+            refreshActions.refreshAll(testDSTree);
             expect(spy).toHaveBeenCalledTimes(1);
         });
     });
