+++ packages/zowe-explorer/src/dataset/actions.ts
@@ -55,21 +55,21 @@ const localize: nls.LocalizeFunc = nls.loadMessageBundle();
  *
  * @param {DataSetTree} datasetProvider
  */
-export async function refreshAll(datasetProvider: IZoweTree<IZoweDatasetTreeNode>) {
-    await Profiles.getInstance().refresh(ZoweExplorerApiRegister.getInstance());
-    datasetProvider.mSessionNodes.forEach(async (sessNode) => {
-        const setting = (await PersistentFilters.getDirectValue("Zowe-Automatic-Validation")) as boolean;
-        if (contextually.isSessionNotFav(sessNode)) {
-            labelRefresh(sessNode);
-            sessNode.children = [];
-            sessNode.dirty = true;
-            refreshTree(sessNode);
-            resetValidationSettings(sessNode, setting);
-            returnIconState(sessNode);
-        }
-    });
-    datasetProvider.refresh();
-}
+// export async function refreshAll(datasetProvider: IZoweTree<IZoweDatasetTreeNode>) {
+//     await Profiles.getInstance().refresh(ZoweExplorerApiRegister.getInstance());
+//     datasetProvider.mSessionNodes.forEach(async (sessNode) => {
+//         const setting = (await PersistentFilters.getDirectValue("Zowe-Automatic-Validation")) as boolean;
+//         if (contextually.isSessionNotFav(sessNode)) {
+//             labelRefresh(sessNode);
+//             sessNode.children = [];
+//             sessNode.dirty = true;
+//             refreshTree(sessNode);
+//             resetValidationSettings(sessNode, setting);
+//             returnIconState(sessNode);
+//         }
+//     });
+//     datasetProvider.refresh();
+// }
 
 /**
  * Allocates a copy of a data set or member
