+++ packages/zowe-explorer/src/shared/refresh.ts
@@ -1,0 +1,40 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the *
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at *
+ * https://www.eclipse.org/legal/epl-v20.html                                      *
+ *                                                                                 *
+ * SPDX-License-Identifier: EPL-2.0                                                *
+ *                                                                                 *
+ * Copyright Contributors to the Zowe Project.                                     *
+ *                                                                                 *
+ */
+
+import { IZoweDatasetTreeNode, IZoweTree, IZoweTreeNode } from "@zowe/zowe-explorer-api";
+import { PersistentFilters } from "../PersistentFilters";
+import { Profiles } from "../Profiles";
+import { refreshTree } from "../utils/ProfilesUtils";
+import { ZoweExplorerApiRegister } from "../ZoweExplorerApiRegister";
+import { resetValidationSettings, returnIconState } from "./actions";
+import { isZoweJobTreeNode, isZoweUSSTreeNode, labelRefresh } from "./utils";
+import * as contextually from "../shared/context";
+
+/**
+ * Refreshes treeView
+ *
+ * @param {DataSetTree} datasetProvider
+ */
+export async function refreshAll(treeProvider: IZoweTree<IZoweTreeNode>) {
+    await Profiles.getInstance().refresh(ZoweExplorerApiRegister.getInstance());
+    treeProvider.mSessionNodes.forEach(async (sessNode) => {
+        const setting = (await PersistentFilters.getDirectValue("Zowe-Automatic-Validation")) as boolean;
+        if (contextually.isSessionNotFav(sessNode)) {
+            labelRefresh(sessNode);
+            sessNode.children = [];
+            sessNode.dirty = true;
+            refreshTree(sessNode);
+            resetValidationSettings(sessNode, setting);
+            returnIconState(sessNode);
+        }
+    });
+    treeProvider.refresh();
+}
